Don't send the cluster struct to the workers
Given that the workers update the story, they will get "self-notified"
when the store triggers on a watch. This caused the worker's buffer to
fill up with notifications that had outdated data.

Instead, we make the workers read from the store when they are poked by
the multi cluster controller.