Further decouple server from plan and controller
This change further decouples the API server from the plan file and the
specifics of how the controller works. It reduces the responsibility of
the API server to just retrieving, validating and storing resources.

The controller, on the other hand, gains the responsibility of creating
an initial plan according to the specification of the cluster that is
created by the user through the API server.

In order to support this, the controller now handles three additional
states:
- "": The empty state is the initial state of all clusters. The
  controller will understand that this cluster was just created, and it
  will transition the cluster to the "planning" state.
- "planning": In this state, the plan is generated by the controller.
  Once the plan has been generated, the cluster is transitioned to the
  "planned" state.
- "planningFailed": The cluster will enter this state if there was an
   error during the "planning" state.

The cluster data structure persisted in the store has also been updated.
Instead of lumping all the state together, the cluster now has a "spec"
field and a "status" field:
- The "spec" field is the specification of the cluster as defined by
  the user. This field can only be written to by the server.
- The "status" field is the current status of the cluster as informed by
  the controller. This field can only be written to by the controller.
  The only caveat right now is that the server still touches one field of
  the cluster status, which is used to communicate "retries".